+++ src/valueObjects/IValued.java
@@ -1,5 +1,7 @@
 package valueObjects;
 
+import java.math.BigDecimal;
+
 /** Represents a calculation on a time series that returns a time series */
 public interface IValued {
 	
@@ -8,5 +10,5 @@ public interface IValued {
 	 * @return The value of the time series ticksBack ticks ago.
 	 * @throws TickOutOfRangeException in case this value is unreachable.
 	 */
-	public int getValue(int ticksBack) throws TickOutOfRangeException;
+	public BigDecimal getValue(int ticksBack) throws TickOutOfRangeException;
 }
